remove things that prevent arm from building
This is predominatly due to instructions which (according to GCC at least)
aren't available in stock arm7di (without the t and the m).

There's still a lot of stuff that needs to be re-written to work on dreamcast
which isn't commented out.  The point of this commit is just to make it build.